+++ bootstrap/github-oidc/main.tf
@@ -29,6 +29,11 @@ provider "aws" {
   }
 }
 
+moved {
+  from = aws_iam_openid_connect_provider.github
+  to   = aws_iam_openid_connect_provider.github[0]
+}
+
 locals {
   default_plan_subjects = [
     "repo:${var.github_org}/${var.github_repo}:pull_request",
@@ -44,19 +49,29 @@ locals {
 
   plan_subject_claims  = length(var.subject_claims_override) > 0 ? var.subject_claims_override : local.default_plan_subjects
   apply_subject_claims = length(var.subject_claims_override) > 0 ? var.subject_claims_override : local.default_apply_subjects
+
+  github_oidc_provider_arn = var.create_github_oidc_provider ? aws_iam_openid_connect_provider.github[0].arn : data.aws_iam_openid_connect_provider.github_existing[0].arn
 }
 
 data "tls_certificate" "github_actions" {
-  url = "https://token.actions.githubusercontent.com"
+  count = var.create_github_oidc_provider ? 1 : 0
+  url   = "https://token.actions.githubusercontent.com"
 }
 
 resource "aws_iam_openid_connect_provider" "github" {
+  count = var.create_github_oidc_provider ? 1 : 0
+
   url            = "https://token.actions.githubusercontent.com"
   client_id_list = ["sts.amazonaws.com"]
   thumbprint_list = distinct(concat(
-    [data.tls_certificate.github_actions.certificates[0].sha1_fingerprint],
+    [data.tls_certificate.github_actions[0].certificates[0].sha1_fingerprint],
     var.additional_thumbprints,
   ))
+}
+
+data "aws_iam_openid_connect_provider" "github_existing" {
+  count = var.create_github_oidc_provider ? 0 : 1
+  arn   = var.existing_github_oidc_provider_arn
 }
 
 data "aws_iam_policy_document" "github_apply_trust" {
@@ -66,7 +81,7 @@ data "aws_iam_policy_document" "github_apply_trust" {
 
     principals {
       type        = "Federated"
-      identifiers = [aws_iam_openid_connect_provider.github.arn]
+      identifiers = [local.github_oidc_provider_arn]
     }
 
     condition {
@@ -90,7 +105,7 @@ data "aws_iam_policy_document" "github_plan_trust" {
 
     principals {
       type        = "Federated"
-      identifiers = [aws_iam_openid_connect_provider.github.arn]
+      identifiers = [local.github_oidc_provider_arn]
     }
 
     condition {